feat: @pollinations/agent SDK (defineAgent) as a shared worker module part
- New pollinations-agent-sdk.ts exports POLLINATIONS_AGENT_SDK_SOURCE: a
  self-contained ESM providing defineAgent(fn) (owns the OpenAI /chat/completions
  + /models envelope, BEE_AUTH_TOKEN auth, SSE framing, and the always-present
  zeroed-token usage + tool_call_counts shape), a keyless gen client bound to the
  injected scoped key, and runTools (the bounded tool loop + built-in tools +
  MCP client).
- deployCommunityWorker uploads the SDK as a SECOND ES-module part on every
  deploy; an index.mjs imports it by relative name (import { defineAgent } from
  "./pollinations-agent.mjs") with no bundler. Verified against CF docs: the raw
  multipart API does no npm resolution but wires module parts by name.
- prompt-agent-template.ts collapses ~435 lines to a thin defineAgent(runTools)
  call that consumes the SDK — dogfoods it and proves it works.
- Tests: existing 4 template tests pass unchanged against the refactor; 4 new
  tests exercise defineAgent directly (string return, object return with
  toolCallCounts, auth rejection before fn runs, error->502).

NOT yet verified end-to-end: that Cloudflare resolves the ./pollinations-agent.mjs
relative import between the two uploaded module parts at the edge. Unit tests
simulate the wiring; needs a staging deploy to confirm.
